Bug #939016 by svendecabooter, dww: Cloned node ID not set via prepopulate.
This functionality was never finished at #654508. Previously, the only
way to know you were inserting a cloned node was if clone.module was
configured to use save-edit, not prepopulate. Now, we use
hook_form_alter() to inject a 'value' form element into the node form
so that $node->clone_from_original_nid is propagated all the way
through node_save() and friends to reach nodeapi('insert').